Stage 1 complete
At the first stage, you need to develop a simple JSON API that always returns the same quiz to be solved. The API should support only two operations: getting the quiz and solving it by passing an answer. Each operation is described in more detail below.
Get the quiz
The quiz has exactly three fields: title (string) text (string) and options (array). To get the quiz, the client sends the GET request to /api/quiz. The server should return the following JSON structure:

{
  "title": "The Java Logo",
  "text": "What is depicted on the Java logo?",
  "options": ["Robot","Tea leaf","Cup of coffee","Bug"]
}
In your API, the names of attributes must be exactly the same (title, text, options), but you can assign any values to them. The quiz should contain four items in the options array. The correct answer must be the third option, but since the indexes start from zero, its index is 2.
Solve the quiz
To solve the quiz, the client need to pass the answer parameter using the POST request to /api/quiz with content as parameter answer and value. This parameter is the index of a chosen option from options array. We suppose that in our service indexes start from zero.

The server should return JSON with two fields: success (true or false) and feedback (just a string). There are two possible responses from the server:

If the passed answer is correct (POST to /api/quiz with content answer=2):
{"success":true,"feedback":"Congratulations, you're right!"}
If the answer is incorrect (e.g., POST to /api/quiz with content answer=1):
{"success":false,"feedback":"Wrong answer! Please, try again."}
You can write any other strings in the feedback field, but the names of the fields and the true/false values must match this example.