Stage 2 complete
Description
At this stage, you will improve the web service to create, get and solve lots of quizzes, not just a single one. All quizzes should be stored in the service's memory, without an external storage.
Create a new quiz
To create a new quiz, the client needs to send a JSON as the request's body via POST to /api/quizzes. The JSON should contain the four fields: title (a string), text (a string), options (an array of strings) and answer (integer index of the correct option). At this moment, all the keys are optional.
Get a quiz by id
To get a quiz by id, the client sends the GET request to /api/quizzes/{id}.
Get all quizzes
To get all existing quizzes in the service, the client sends the GET request to /api/quizzes.
Solving a quiz
To solve the quiz, the client sends a POST request to /api/quizzes/{id}/solve and passes the answer parameter in the content. This parameter is the index of a chosen option from options array. As before, it starts from zero.