Stage 4 complete
In the implementation of this step , I tried converting QUIZ objects obtained as a result of a POST request to QUIZ_ENTITY objects for placement in the database. And vice versa.

Description
At this stage, you will permanently store the data in a database, so that after restarting the service you will not lose all quizzes created by the users. You don't need to change the API of your service at this stage.
If that doesn't feel like enough, you can google other articles on the same topics (Spring Data, Repositories, JPA, H2).
We recommend you use the H2 database in the disk-based storage mode (not in-memory).

You can use any tables in your database to complete this stage. The main thing is that when you restart the service, quizzes should not be lost. Our tests will create and get them via the API developed at the previous stages.